Setting Email Authentication | req0-2 | S1P3151005-40

+++ back/src/main/java/com/ssafy/model/Room.java
@@ -35,7 +35,7 @@ class Room {
     private Boolean completeFlag;
 
     @Column(name = "USER_COUNT", nullable = false)
-    private Integer userCount;
+    private Short userCount;
 
     @OneToMany(mappedBy = "room")
     private List<Sprint> sprints = new ArrayList<Sprint>();
